refactor: 인수 테스트 관련 기능, 추상 클래스로 이동
